Initialise firebase in separate file to avoid require cycles
