+++ capsulert/App.js
@@ -1,5 +1,3 @@
-import { initializeApp } from "firebase/app";
-import { getAuth } from "@firebase/auth";
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import SignInPage from "./components/SignInPage";
@@ -8,19 +6,6 @@ import MyMedical from "./components/MyMedical";
 import { UserProvider } from "./contexts/User";
 
 const Stack = createNativeStackNavigator();
-
-const firebaseConfig = {
-  apiKey: "AIzaSyDc6-4VudZ4SMt5dm2iS7CoTH8WN6nme0s",
-  authDomain: "capsulert.firebaseapp.com",
-  projectId: "capsulert",
-  storageBucket: "capsulert.appspot.com",
-  messagingSenderId: "775759460960",
-  appId: "1:775759460960:web:3eb3a47391236e87616105",
-  measurementId: "G-33TQPKC9KF",
-};
-
-const app = initializeApp(firebaseConfig);
-export const auth = getAuth();
 
 export default function App() {
   return (
